rename to externals.d.ts per suggestion
